requirements.txt
beautifulsoup4==4.9.1
Brotli==1.0.7
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dash==1.14.0
dash-bootstrap-components==0.10.3
dash-core-components==1.10.2
dash-daq==0.5.0
dash-html-components==1.0.3
dash-renderer==1.6.0
dash-table==4.9.0
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
html5lib==1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.2
MarkupSafe==1.1.1
numpy==1.19.1
pandas==1.1.0
plotly==4.9.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
retrying==1.3.3
scraperapi-sdk==0.2.2
six==1.15.0
soupsieve==2.0.1
Unidecode==1.1.1
urllib3==1.25.10
webencodings==0.5.1
Werkzeug==1.0.1
xlrd==1.2.0
